cleanup and move stuff
